+++ package.json
@@ -29,6 +29,7 @@
     "@typescript-eslint/eslint-plugin": "^4.12.0",
     "@typescript-eslint/parser": "^4.12.0",
     "postcss": "^8.1.6",
+    "react-device-detect": "^1.17.0",
     "sass": "^1.26.11",
     "ts-node": "^9.1.1",
     "typescript": "^4.0.5"
