--- src/index.d.ts
@@ -1,8 +1,0 @@
-declare module '*.scss' {
-  interface IClassNames {
-    [className: string]: string;
-  }
-
-  const classNames: IClassNames;
-  export = classNames;
-}
