+++ src/pages/index.scss
