--- src/pages/member.test.tsx
@@ -1,10 +1,0 @@
-import { render, screen } from '@testing-library/react';
-
-import Member from './member';
-
-describe('Member', () => {
-  test('check Member behavior', () => {
-    render(<Member />);
-    expect(screen.getByText('ユニット', { exact: false })).toBeInTheDocument;
-  });
-});
