--- src/pages/road.test.tsx
@@ -1,10 +1,0 @@
-import { render, screen } from '@testing-library/react';
-
-import Road from './road';
-
-describe('Road', () => {
-  test('check Road behavior', () => {
-    render(<Road />);
-    expect(screen.getByText('歩んできた道のり', { exact: false })).toBeInTheDocument;
-  });
-});
